Extracted SpectatorInfoPane from TutorialPane, some more TutorialPane refactoring
