+++ src/Frontend/Panes/Game/StyledTutorialPane.tsx
@@ -1,0 +1,38 @@
+import styled from 'styled-components';
+import dfstyles from '../../Styles/dfstyles';
+
+export const StyledTutorialPane = styled.div`
+  position: absolute;
+  top: 0;
+  left: 0;
+
+  background: ${dfstyles.colors.backgroundlighter};
+  color: ${dfstyles.colors.text};
+  padding: 8px;
+  border-bottom: 1px solid ${dfstyles.colors.border};
+  border-right: 1px solid ${dfstyles.colors.border};
+
+  width: 24em;
+  height: fit-content;
+
+  z-index: 10;
+
+  & .tutintro {
+    & > div:last-child {
+      display: flex;
+      flex-direction: row;
+      justify-content: space-around;
+      margin-top: 1em;
+    }
+  }
+
+  & .tutzoom,
+  & .tutalmost {
+    & > div:last-child {
+      display: flex;
+      flex-direction: row;
+      justify-content: flex-end;
+      margin-top: 1em;
+    }
+  }
+`;
